Refactor: Separa CollectorAlbumDto en archivo independiente
- Crea archivo CollectorAlbumDto.kt siguiendo el patrón del proyecto
- Mantiene una clase por archivo para mejorar legibilidad
- Atiende comentario de revisión sobre buenas prácticas de DTOs
